requirement.txt

blessings==1.7
bpython==0.19
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.4.5.2
chardet==3.0.4
curtsies==0.3.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.5
django-crontab==0.7.1
django-import-export==2.1.0
django-jet==1.0.8
django-jsonfield==1.4.0
django-simple-history==2.11.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
feedparser==5.2.1
firebase-admin==4.3.0
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.4
google-cloud-core==1.3.0
google-cloud-firestore==1.8.1
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
greenlet==0.4.16
grpcio==1.30.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
msgpack==1.0.0
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.2.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
rsa==4.6
six==1.15.0
tablib==1.1.0
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
